Implementación código de ejemplo, consumo web service Zona A

+++ Imed_Api/Models/Licencias/RecuperoZonaA.cs
@@ -3,12 +3,12 @@
     public class RecuperoZonaA
     {
         public string CodigoOperador { get; set; }
-        public long RutEmpleador { get; set; }
+        public int RutEmpleador { get; set; }
         public string DvEmpleador { get; set; }
         public string IdUnidadrrhh { get; set; }
         public string Clave { get; set; }
 
-        public long NumLicencia { get; set; }
+        public int NumLicencia { get; set; }
 
         public string DigLicencia { get; set; }
     }
